256 - colors

+++ src/main/java/dotBlueShoes/fogger/config/PropertyFogColor.java
@@ -3,9 +3,9 @@ package dotBlueShoes.fogger.config;
 public class PropertyFogColor {
 
 	public String name, _comment;
-	public float r, g, b;
+	public int r, g, b;
 
-	public PropertyFogColor(String name, String _comment, float r, float g, float b) {
+	public PropertyFogColor(String name, String _comment, int r, int g, int b) {
 		this._comment = _comment;
 		this.name = name;
 		this.r = r;
